refactor: rename S3Image -> ImageFile
