+++ apps/storage/src/ImageFile.ts
