+++ apps/storage/src/index.ts
@@ -4,5 +4,5 @@ export * from "./get.js";
 export * from "./getPublicUrl.js";
 export * from "./upload.js";
 export * from "./signed-url.js";
+export * from "./ImageFile.js";
 export * from "./s3.js";
-export * from "./S3Image.js";
